Ensure exceptions in handlers don't propagate any further
Used TPL dataflow to effectively queue up incoming messages.

Then had to change how stopping the bot works. It now must wait for the message handling to be finished before it can stop.
This then required expanding IMessageTransport a bit so that incoming messages could be stopped, while response messages can still be sent.
Did some member renaming of the interface, and further solidified the design where transports only should expect one bot.

With this change, the tests needed some updates to move the asserts to after the bot is stopped.

Also threw in a minor refactoring of the message handlers used in tests.